Software/requirements.txt
pyserial
flet